# application/services/custom_services/servicetypes/src/models/entitymodels/servicestype.ts
export interface servicestype 
{
   created_date: { type: Date },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date },
   name: { type: String },
   description: { type: String }
}